refactor: Skill 클래스 주 도메인으로 패키징
